Refactor watchOS views into separate files
Resolves #6

Split monolithic ContentView (330 lines) into focused, maintainable SwiftUI views:
- RootView: Navigation coordinator owning GameViewModel
- MainMenuView: Title, high score, play button
- GamePlayView: Timer ring, command display, gesture buttons
- GameOverView: Score display, tap to continue
- CountdownView: 3-2-1-GO countdown
- SpeedUpMessageView: Speed-up interstitial
- ViewModifiers/JiggleEffect: Reusable animation modifier

All views use @ObservedObject pattern with RootView owning @StateObject.
Each view includes SwiftUI preview. Zero behavioral changes.
GameViewModel and all tests remain unchanged and pass.
