+++ WristBop/WristBop Watch App/WristBopApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct WristBop_Watch_AppApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            RootView()
         }
     }
 }
